--- src/components/MyPage/SavedFeeds/SavedFeeds.module.scss
@@ -1,34 +1,0 @@
-@use "@/styles/globals.scss" as *;
-
-.cardContainer {
-  margin-top: 20px;
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  row-gap: 40px;
-  column-gap: 16px;
-
-  @include Size("mobile") {
-    grid-template-columns: repeat(2, 1fr);
-    row-gap: 20px;
-    column-gap: 12px;
-    margin-top: 68px;
-  }
-
-  @include Size("tablet") {
-    grid-template-columns: repeat(3, 1fr);
-    row-gap: 20px;
-    column-gap: 16px;
-  }
-}
-
-.noResult {
-  @include Sub2;
-  color: $gray50;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 100px 0;
-  gap: 24px;
-}
